Removed Code that cleared key Input
For 3 different opcodes there was a line that cleared the input of a key
after hit has been used for an opcode. The intention was to avoid
opcodes reading the same key input accidentally. This is not necessary
as the JS code already clears the key immediately when let go.