Removal of drawCanvas()
XorPixel now draws pixels directly instead of modifying the array and
using drawCanvas to redraw the full canvas. This was quite efficient.
Though it allowed the skipping of drawing when a pixel was disabled
which reduced flickering.

This change makes flickering more apparent but will lead to less work
each loop.